requirements.txt
torch
torchvision
argparse
# shutil
# warnings
# datetime
opencv-python
numpy
albumentations
pillow